requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Backtesting
vectorbt>=0.24.0
empyrical>=0.5.5

# Data fetching
aiohttp>=3.8.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Utils
pyyaml>=6.0
numba>=0.56.0
tqdm>=4.64.0

# Development
pytest>=7.2.0
black>=22.10.0
flake8>=5.0.0
mypy>=0.990